requirements.txt
ipinfo==4.2.0
invoke==1.5.0
requests==2.25.1
GitPython==3.1.17
uwsgi==2.0.19.1
